chatbot/requirements.txt
Flask
flask_cors
python-dotenv
requests
sentence-transformers
torch
gunicorn

